fix warnings with older compiler
git-svn-id: https://svn.salilab.org/imp/trunk@12202 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/domino/include/domino_macros.h
@@ -82,7 +82,7 @@
     }                                                                   \
   };                                                                    \
   struct Name##Next {                                                   \
-    double operator()(int pos, const Assignment& state,                 \
+    int operator()(int pos, const Assignment& state,                    \
                       const Ints &set) const {                          \
       next;                                                             \
     }                                                                   \
